+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## unreleased
 
-### pack, serve
+### serve
+
+* Fixed `enact serve` fails to open another port when the default port is busy.
+
+### pack
 
 * Fixed `core-js` version to `3.22.8` for compatibility.
 * Added `ENACT_PACK_ISOMORPHIC` as a global variable to use `hydrateRoot` instead of `createRoot` from app when isomorphic build.